requirements.txt
# Minimal, pinned dependencies for CPU or single-GPU
# If you have CUDA, install the matching torch build from https://pytorch.org/get-started/locally/

torch>=2.2,<3.0
numpy>=1.24,<3.0
pandas>=2.0,<3.0
pyyaml>=6.0,<7.0
tqdm>=4.66,<5.0
biopython>=1.83,<2.0
scikit-learn>=1.3,<2.0
# FAIR-ESM for ESM2 models
esm>=3.0.0
